deps/v8/test/inspector/debugger/pause-inside-blackboxed-optimized-expected.txt

Checks pause inside blackboxed optimized function call.

Running test: test
bar (test.js:2:4)
foo (framework.js:2:15)
(anonymous) (expr.js:1:12)
